refactor: fix typo depricated deprecated
