+++ src/gaussian/deprecated.rs
